some warnings with format

+++ Sources/DatabaseCollectionOutlineView/SidebarBadgeCell.m
@@ -95,7 +95,7 @@
 - (CGFloat)drawBadge:(NSRect)cellFrame
 {
     // Setup Badge String and Size
-    NSString *badge = [[NSString alloc] initWithFormat:@"%ld", _badgeCount];
+    NSString *badge = [[NSString alloc] initWithFormat:@"%ld", (long)_badgeCount];
     NSSize badgeNumSize = [badge sizeWithAttributes:nil];
     NSFont *badgeFont = [self font];
